quick fix

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -8,7 +8,7 @@
 * Return: 1 if it worked, or -1 if an error occurred
 */
 
-int set_bit(unsigned long int *n, unsigned long int index)
+int set_bit(unsigned long int *n, unsigned int index)
 {
 	if ((index >= sizeof(long) * 8))
 	{
